# Q2k interleaving implementation - x86/x64 SIMD