@slaren 2025-07-04
I tested this on a 13900k with gcc 13 and clang 19, but the improvement is not very significant. Repacking has a significant cost, since it increases load time and prevents usage of mmap, and as it is, I find this very hard to justify for AVX2. It may make sense for AVX512, but I cannot test that.


<details>

GCC-13:

| Model           |   Threads | Test   |   t/s master |   t/s q2k_interleaving_implementation |   Speedup |
|:----------------|----------:|:-------|-------------:|--------------------------------------:|----------:|
| llama 7B Q2_K_M |         8 | pp64   |        47.56 |                                 48.55 |      1.02 |
| llama 7B Q2_K_M |         8 | tg32   |        19.92 |                                 19.38 |      0.97 |
| llama 7B Q2_K_M |        16 | pp64   |        63.04 |                                 60.08 |      0.95 |
| llama 7B Q2_K_M |        16 | tg32   |        21.22 |                                 20.44 |      0.96 |
| llama 7B Q2_K_M |        24 | pp64   |        68.39 |                                 68.07 |      1.00 |
| llama 7B Q2_K_M |        24 | tg32   |        19.72 |                                 19.76 |      1.00 |
| llama 7B Q2_K_M |        32 | pp64   |        71.18 |                                 71.62 |      1.01 |
| llama 7B Q2_K_M |        32 | tg32   |        17.87 |                                 17.51 |      0.98 |


Clang-19:
| Model           |   Threads | Test   |   t/s master |   t/s q2k_interleaving_implementation |   Speedup |
|:----------------|----------:|:-------|-------------:|--------------------------------------:|----------:|
| llama 7B Q2_K_M |         8 | pp64   |        48.28 |                                 52.27 |      1.08 |
| llama 7B Q2_K_M |         8 | tg32   |        20.78 |                                 19.08 |      0.92 |
| llama 7B Q2_K_M |        16 | pp64   |        65.23 |                                 61.42 |      0.94 |
| llama 7B Q2_K_M |        16 | tg32   |        20.94 |                                 19.79 |      0.95 |
| llama 7B Q2_K_M |        24 | pp64   |        69.69 |                                 71.17 |      1.02 |
| llama 7B Q2_K_M |        24 | tg32   |        19.90 |                                 19.59 |      0.98 |
| llama 7B Q2_K_M |        32 | pp64   |        71.04 |                                 75.26 |      1.06 |
| llama 7B Q2_K_M |        32 | tg32   |        16.91 |                                 17.30 |      1.02 |

</details>